+++ source/isaaclab/isaaclab/terrains/terrain_importer.py
@@ -5,12 +5,11 @@
 
 from __future__ import annotations
 
+import logging
 import numpy as np
 import torch
 import trimesh
 from typing import TYPE_CHECKING
-
-import omni.log
 
 import isaaclab.sim as sim_utils
 from isaaclab.markers import VisualizationMarkers
@@ -20,6 +19,9 @@ from .utils import create_prim_from_mesh
 
 if TYPE_CHECKING:
     from .terrain_importer_cfg import TerrainImporterCfg
+
+# import logger
+logger = logging.getLogger(__name__)
 
 
 class TerrainImporter:
@@ -211,7 +213,7 @@ class TerrainImporter:
             if "diffuse_color" in material:
                 color = material["diffuse_color"]
             else:
-                omni.log.warn(
+                logger.warning(
                     "Visual material specified for ground plane but no diffuse color found."
                     " Using default color: (0.0, 0.0, 0.0)"
                 )
@@ -376,7 +378,7 @@ class TerrainImporter:
         .. deprecated:: v2.1.0
             The `warp_meshes` attribute is deprecated. It is no longer stored inside the class.
         """
-        omni.log.warn(
+        logger.warning(
             "The `warp_meshes` attribute is deprecated. It is no longer stored inside the `TerrainImporter` class."
             " Returning an empty dictionary."
         )
@@ -389,7 +391,7 @@ class TerrainImporter:
         .. deprecated:: v2.1.0
             The `meshes` attribute is deprecated. It is no longer stored inside the class.
         """
-        omni.log.warn(
+        logger.warning(
             "The `meshes` attribute is deprecated. It is no longer stored inside the `TerrainImporter` class."
             " Returning an empty dictionary."
         )